+++ Runtime/IOneFrameEvent.cs
@@ -1,0 +1,7 @@
+﻿namespace ED.DOTS.OneFrameEvents
+{
+    public interface IOneFrameEvent
+    {
+        
+    }
+}