+++ Runtime/OneFrameEventSimulationSystemGroup.cs
@@ -1,0 +1,10 @@
+﻿using Unity.Entities;
+
+namespace ED.DOTS.OneFrameEvents
+{
+    [UpdateInGroup(typeof(SimulationSystemGroup), OrderLast = true)]
+    [UpdateAfter(typeof(LateSimulationSystemGroup))]
+    [UpdateBefore(typeof(EndSimulationEntityCommandBufferSystem))]
+    [WorldSystemFilter(WorldSystemFilterFlags.Default | WorldSystemFilterFlags.ThinClientSimulation)]
+    public partial class OneFrameEventSimulationSystemGroup : ComponentSystemGroup { }
+}